src/requirements.txt
experta==1.9.4
python-telegram-bot==20.7
pgmpy==0.1.25
pandas==2.1.4
numpy==1.25.2